requirements.txt
asgiref==3.7.2
Django==4.2.2
django-filter==23.2
djangorestframework==3.14.0
iniconfig==2.0.0
model-bakery==1.12.0
packaging==23.1
pluggy==1.0.0
psycopg2-binary==2.9.6
pytest==7.3.1
pytest-django==4.5.2
pytz==2023.3
sqlparse==0.4.4
